Catch 404 in paging next, add exception hierarchy
