+++ tekore/client/api/user.py
@@ -1,6 +1,7 @@
 from tekore.model import PublicUser, PrivateUser
+from tekore.client.base import SpotifyBase
+from tekore.client.decor import send_and_process
 from tekore.client.process import single
-from tekore.client.base import SpotifyBase, send_and_process
 
 
 class SpotifyUser(SpotifyBase):
