+++ tekore/client/paging.py
@@ -1,8 +1,10 @@
 from typing import Generator, Optional
 
-from tekore.model.paging import Paging, OffsetPaging
-from tekore.client.base import send_and_process, SpotifyBase
 from tekore.serialise import SerialisableDataclass
+from tekore.client.base import SpotifyBase
+from tekore.model.paging import Paging, OffsetPaging
+from tekore.client.decor import send_and_process
+from tekore.client.decor.error import NotFound
 
 
 def parse_paging_result(result):
@@ -39,15 +41,21 @@ class SpotifyPaging(SpotifyBase):
         if page.next is None:
             return
 
-        next_set = self._get_paging_result(page.next)
-        return type(page)(**next_set)
+        try:
+            next_set = self._get_paging_result(page.next)
+            return type(page)(**next_set)
+        except NotFound:
+            return
 
     async def _async_next(self, page: Paging) -> Optional[Paging]:
         if page.next is None:
             return
 
-        next_set = await self._get_paging_result(page.next)
-        return type(page)(**next_set)
+        try:
+            next_set = await self._get_paging_result(page.next)
+            return type(page)(**next_set)
+        except NotFound:
+            return
 
     def previous(self, page: OffsetPaging) -> Optional[OffsetPaging]:
         """
@@ -102,16 +110,14 @@ class SpotifyPaging(SpotifyBase):
             return self._sync_all_pages(page)
 
     def _sync_all_pages(self, page: Paging):
-        yield page
-        while page.next is not None:
-            page = self.next(page)
+        while page is not None:
             yield page
+            page = self.next(page)
 
     async def _async_all_pages(self, page: Paging):
-        yield page
-        while page.next is not None:
-            page = await self._async_next(page)
+        while page is not None:
             yield page
+            page = await self._async_next(page)
 
     def all_items(
             self,
